rounded out agent_groups docs and embedded bulk ops into the relevent methods #25

+++ tenable/io/agent_groups.py
@@ -1,44 +1,70 @@
+'''
+agent_groups
+============
+
+The following methods allow for interaction into the Tenable.io 
+`agent groups <https://cloud.tenable.com/api#/resources/agent-groups>`_ 
+API endpoints.
+
+Methods available on ``tio.agent_groups``:
+
+.. rst-class:: hide-signature
+.. autoclass:: AgentGroupsAPI
+
+    .. automethod:: add_agent
+    .. automethod:: configure
+    .. automethod:: create
+    .. automethod:: delete
+    .. automethod:: delete_agent
+    .. automethod:: details
+    .. automethod:: task_status
+'''
 from .base import TIOEndpoint
 
 class AgentGroupsAPI(TIOEndpoint):
     def add_agent(self, group_id, agent_id, scanner_id=1):
         '''
+        Adds an agent or multiple agents to the agent group specified.
+
         `agent-groups: add-agent <https://cloud.tenable.com/api#/resources/agent-groups/add-agent>`_
 
         Args:
             group_id (int): The id of the group
-            agent_id (int): The id of the agent
+            agent_id (int or list): The id of the agent
             scanner_id (int, optional): The id of the scanner
 
         Returns:
-            None: Agent added successfully
+            None: Single Agent added successfully
+            dict: Task resource if multiple Agents were added.
+
+        Examples:
+            Adding a singular agent:
+
+            >>> tio.agent_groups.add_agent(1, 1)
+
+            Adding multiple agents:
+
+            >>> tio.agent_groups.add_agent(1, [1, 2, 3])
         '''
-        self._api.put('scanners/{}/agent-groups/{}/agents/{}'.format(
-            self._check('scanner_id', scanner_id, int),
-            self._check('group_id', group_id, int),
-            self._check('agent_id', agent_id, int)
-        ))
-
-    def bulk_add_agent(self, group_id, agent_ids, scanner_id=1):
-        '''
-        `bulk-operations: bulk-add-agent <https://cloud.tenable.com/api#/resources/bulk-operations/bulk-add-agent>`_
-
-        Args:
-            group_id (int): The id of the group
-            agent_id (int): The id of the agent
-            scanner_id (int, optional): The id of the scanner
-
-        Returns:
-            dict: Task resource
-        '''
-        self._api.post('scanners/{}/agent-groups/{}/agents/_bulk/add'.format(
-            self._check('scanner_id', scanner_id, int),
-            self._check('group_id', group_id, int)),
-            json={'items': self._check('agent_ids', agent_ids, list)})
-
+        if isinstance(agent_id, list):
+            # if the agent_id field is a list, then perform the bulk operation.
+            return self._api.post(
+                'scanners/{}/agent-groups/{}/agents/_bulk/add'.format(
+                    self._check('scanner_id', scanner_id, int),
+                    self._check('group_id', group_id, int)),
+                json={'items': self._check('agent_id', agent_id, list)}).json()
+        else:
+            # if the agent_id isn't a list, we will perform a singular add.
+            self._api.put('scanners/{}/agent-groups/{}/agents/{}'.format(
+                self._check('scanner_id', scanner_id, int),
+                self._check('group_id', group_id, int),
+                self._check('agent_id', agent_id, int)
+            ))
 
     def configure(self, group_id, name, scanner_id=1):
         '''
+        Renames an existing agent group.
+
         `agent-groups: configure <https://cloud.tenable.com/api#/resources/agent-groups/configure>`_
 
         Args:
@@ -48,6 +74,9 @@ class AgentGroupsAPI(TIOEndpoint):
 
         Returns:
             None: Agent group updated successfully
+
+        Examples:
+            >>> tio.agent_groups.configure(1, 'New Name')
         '''
         self._api.put('scanners/{}/agent-groups/{}'.format(
             self._check('scanner_id', scanner_id, int),
@@ -56,6 +85,8 @@ class AgentGroupsAPI(TIOEndpoint):
 
     def create(self, name, scanner_id=1):
         '''
+        Creates a new agent group.
+
         `agent-groups: create <https://cloud.tenable.com/api#/resources/agent-groups/create>`_
 
         Args:
@@ -65,7 +96,10 @@ class AgentGroupsAPI(TIOEndpoint):
 
         Returns:
             dict: 
-                The dictionary object representing the newly minted agent group 
+                The dictionary object representing the newly minted agent group
+
+        Examples:
+            >>> group = tio.agent_groups.create('New Agent Group')
         '''
         return self._api.post(
             'scanners/{}/agent-groups'.format(
@@ -74,6 +108,8 @@ class AgentGroupsAPI(TIOEndpoint):
 
     def delete(self, group_id, scanner_id=1):
         '''
+        Delete an agent group.
+
         `agent-groups: delete <https://cloud.tenable.com/api#/resources/agent-groups/delete>`_
 
         Args:
@@ -82,6 +118,9 @@ class AgentGroupsAPI(TIOEndpoint):
 
         Returns:
             None: Agent group deleted successfully
+
+        Examples:
+            >>> tio.agent_groups.delete(1)
         '''
         self._api.delete('scanners/{}/agent-groups/{}'.format(
             self._check('scanner_id', scanner_id, int),
@@ -90,41 +129,47 @@ class AgentGroupsAPI(TIOEndpoint):
 
     def delete_agent(self, group_id, agent_id, scanner_id=1):
         '''
+        Delete one or many agents from an agent group.
+
         `agent-groups: delete-agent <https://cloud.tenable.com/api#/resources/agent-groups/delete-agent>`_
 
         Args:
             group_id (int): The id of the agent group to remove the agent from
-            agent_id (int): The id of the agent to be removed
+            agent_id (int ot list): The id of the agent to be removed
             scanner_id (int, optional): The id of the scanner
 
         Returns:
-            None: The agent was deleted from the group successfully.
-        '''
-        self._api.delete('scanners/{}/agent-groups/{}/agents/{}'.format(
-            self._check('scanner_id', scanner_id, int),
-            self._check('group_id', group_id, int),
-            self._check('agent_id', agent_id, int)
-        ))
+            None: A singular agent was deleted from the group successfully.
+            dict: Task resource if multiple agents were requested to be deleted.
 
-    def bulk_delete_agent(self, group_id, agent_ids, scanner_id=1):
-        '''
-        `bulk-operations: bulk-add-agent <https://cloud.tenable.com/api#/resources/bulk-operations/bulk-remove-agent>`_
+        Examples:
+            Delete a singular agent from an agent group:
 
-        Args:
-            group_id (int): The id of the group
-            agent_id (int): The id of the agent
-            scanner_id (int, optional): The id of the scanner
+            >>> tio.agent_groups.delete(1, 1)
 
-        Returns:
-            dict: Task resource
+            Delete multiple agents from an agent group:
+
+            >>> tio.agent_groups.delete(1, [1, 2, 3])
         '''
-        self._api.post('scanners/{}/agent-groups/{}/agents/_bulk/add'.format(
-            self._check('scanner_id', scanner_id, int),
-            self._check('group_id', group_id, int)),
-            json={'items': self._check('agent_ids', agent_ids, list)})
+        if isinstance(agent_id, list):
+            # if multiple agent ids were requested to be deleted, then we will
+            # call the bulk deletion API.
+            return self._api.post(
+                'scanners/{}/agent-groups/{}/agents/_bulk/add'.format(
+                    self._check('scanner_id', scanner_id, int),
+                    self._check('group_id', group_id, int)),
+                json={'items': self._check('agent_ids', agent_id, list)}).json()
+        else:       
+            self._api.delete('scanners/{}/agent-groups/{}/agents/{}'.format(
+                self._check('scanner_id', scanner_id, int),
+                self._check('group_id', group_id, int),
+                self._check('agent_id', agent_id, int)
+            ))
 
     def details(self, group_id, scanner_id=1):
         '''
+        Retrieve the details about the specified agent group.
+
         `agent-groups: details <https://cloud.tenable.com/api#/resources/agent-groups/details>`_
 
         Args:
@@ -134,6 +179,10 @@ class AgentGroupsAPI(TIOEndpoint):
         Returns:
             dict:
                 The dictionary object representing the requested agent group
+
+        Examples:
+            >>> group = tio.agent_groups.details(1)
+            >>> pprint(group)
         '''
         return self._api.get(
             'scanners/{}/agent-groups/{}'.format(
@@ -141,13 +190,15 @@ class AgentGroupsAPI(TIOEndpoint):
                 self._check('group_id', group_id, int)
             )).json()
 
-    def bulk_status(self, group_id, task_uuid, scanner_id=1):
+    def task_status(self, group_id, task_uuid, scanner_id=1):
         '''
+        Retrieves the current status of a bulk task.
+
         `bulk-operations: bulk-agent-group-status <https://cloud.tenable.com/api#/resources/bulk-operations/bulk-agent-group-status>`_
 
         Args:
             group_id (int): The id of the group
-            task_uuid (str): The id of the agent
+            task_uuid (str): The id of the task
             scanner_id (int, optional): The id of the scanner
 
         Returns:
